Move file mgmt state to RequirementForm
