+++ src/app/(protected)/org/[orgId]/[projectId]/requirements/[requirementSlug]/components/RequirementForm.tsx
@@ -6,7 +6,15 @@ import {
     Upload,
     Wand,
 } from 'lucide-react';
-import { ChangeEvent, useRef } from 'react';
+import {
+    ChangeEvent,
+    Dispatch,
+    SetStateAction,
+    useEffect,
+    useMemo,
+    useRef,
+    useState,
+} from 'react';
 
 import { Button } from '@/components/ui/button';
 import { Card } from '@/components/ui/card';
@@ -18,34 +26,40 @@ import {
     SelectTrigger,
     SelectValue,
 } from '@/components/ui/select';
+import {
+    useUpdateExternalDocumentGumloopName,
+    useUploadExternalDocument,
+} from '@/hooks/mutations/useExternalDocumentsMutations';
+import { useExternalDocumentsByOrg } from '@/hooks/queries/useExternalDocuments';
+import { useChunkr } from '@/hooks/useChunkr';
+import { useGumloop } from '@/hooks/useGumloop';
+import { TaskResponse, TaskStatus } from '@/lib/services/chunkr';
 
 interface RequirementFormProps {
+    organizationId: string;
     requirement: {
         id: string;
         name?: string;
     };
     reqText: string;
-    setReqText: (text: string) => void;
+    setReqText: Dispatch<SetStateAction<string>>;
     systemName: string;
-    setSystemName: (name: string) => void;
+    setSystemName: Dispatch<SetStateAction<string>>;
     objective: string;
-    setObjective: (objective: string) => void;
+    setObjective: Dispatch<SetStateAction<string>>;
     isReasoning: boolean;
-    setIsReasoning: (value: boolean) => void;
+    setIsReasoning: Dispatch<SetStateAction<boolean>>;
     isAnalysing: boolean;
     handleAnalyze: () => void;
     missingReqError: string;
     missingFilesError: string;
-    isUploading: boolean;
-    uploadButtonText: string;
-    handleFileUpload: (e: ChangeEvent<HTMLInputElement>) => void;
-    existingDocs?: Array<{ id: string; name: string }>;
-    existingDocsValue: string;
-    handleExistingDocSelect: (docName: string) => void;
+    setMissingFilesError: Dispatch<SetStateAction<string>>;
     selectedFiles: { [key: string]: string };
+    setSelectedFiles: Dispatch<SetStateAction<{ [key: string]: string }>>;
 }
 
 export function RequirementForm({
+    organizationId,
     requirement,
     reqText,
     setReqText,
@@ -59,19 +73,224 @@ export function RequirementForm({
     handleAnalyze,
     missingReqError,
     missingFilesError,
-    isUploading,
-    uploadButtonText,
-    handleFileUpload,
-    existingDocs,
-    existingDocsValue,
-    handleExistingDocSelect,
+    setMissingFilesError,
     selectedFiles,
+    setSelectedFiles,
 }: RequirementFormProps) {
     const fileInputRef = useRef<HTMLInputElement>(null);
 
     const handleInputChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
         setReqText(e.target.value);
     };
+
+    const [isUploading, setIsUploading] = useState(false);
+    const [processingPdfFiles, setProcessingPdfFiles] = useState<
+        { name: string; supabaseId: string }[]
+    >([]);
+
+    const uploadFileToSupabase = useUploadExternalDocument();
+    const updateGumloopName = useUpdateExternalDocumentGumloopName();
+
+    const { uploadFiles } = useGumloop();
+    const { startOcrTask, getTaskStatuses } = useChunkr();
+    const [ocrPipelineTaskIds, setOcrPipelineTaskIds] = useState<string[]>([]);
+    const taskStatusQueries = getTaskStatuses(ocrPipelineTaskIds);
+
+    const handleFileUpload = async (e: ChangeEvent<HTMLInputElement>) => {
+        if (!e.target.files?.length) return;
+
+        try {
+            const files = Array.from(e.target.files);
+            setIsUploading(true);
+            setMissingFilesError('');
+
+            // upload to Supabase
+            const uploadPromises = files.map((file) =>
+                uploadFileToSupabase.mutateAsync({
+                    file,
+                    orgId: organizationId,
+                }),
+            );
+            const supabaseUploads = await Promise.all(uploadPromises);
+            console.log('Uploaded files to Supabase');
+
+            const nonPdfFiles = [];
+            const pdfFiles = [];
+
+            for (let i = 0; i < files.length; i++) {
+                const file = files[i];
+                const supabaseId = supabaseUploads[i].id;
+                if (file.name.toLowerCase().endsWith('.pdf')) {
+                    pdfFiles.push({
+                        file,
+                        supabaseId,
+                        name: file.name,
+                    });
+                } else {
+                    nonPdfFiles.push({
+                        file,
+                        supabaseId,
+                        name: file.name,
+                    });
+                }
+            }
+
+            // upload to Gumloop
+            if (nonPdfFiles.length > 0) {
+                await uploadFiles(nonPdfFiles.map((file) => file.file));
+                console.log('Uploaded files to Gumloop');
+
+                // Add non-PDF files directly to selectedFiles
+                nonPdfFiles.forEach((file) => {
+                    setSelectedFiles((prev) => ({
+                        ...prev,
+                        [file.name]: file.name,
+                    }));
+                });
+
+                const gumloopNamePromises = nonPdfFiles.map((file) => {
+                    updateGumloopName.mutateAsync({
+                        documentId: file.supabaseId,
+                        gumloopName: file.name,
+                        orgId: organizationId,
+                    });
+                });
+                await Promise.all(gumloopNamePromises);
+                console.log('Updated gumloop names in Supabase');
+            }
+
+            if (pdfFiles.length == 0) {
+                setIsUploading(false);
+                return;
+            }
+
+            setProcessingPdfFiles(pdfFiles);
+
+            const taskIds = await startOcrTask(
+                pdfFiles.map((file) => file.file),
+            );
+            setOcrPipelineTaskIds(taskIds);
+        } catch (error) {
+            setIsUploading(false);
+            console.error('Failed to upload files:', error);
+        }
+    };
+
+    const { data: existingDocs } = useExternalDocumentsByOrg(organizationId);
+    const existingDocsNameMap = useMemo(() => {
+        if (!existingDocs) return {};
+        return existingDocs.reduce(
+            (acc, doc) => {
+                acc[doc.name] = doc.gumloop_name || doc.name;
+                return acc;
+            },
+            {} as { [key: string]: string },
+        );
+    }, [existingDocs]);
+    const [existingDocsValue, setExistingDocsValue] = useState<string>('');
+
+    const handleExistingDocSelect = (docName: string) => {
+        setMissingFilesError('');
+        setSelectedFiles((prev) => ({
+            ...prev,
+            [existingDocsNameMap[docName]]: docName,
+        }));
+        setExistingDocsValue('');
+    };
+
+    // set the selectedFiles when the pipeline run is completed
+    useEffect(() => {
+        // Check if taskStatuses array is empty or any queries are still loading
+        if (!taskStatusQueries.length) {
+            return;
+        }
+
+        // Check if any task failed
+        if (
+            taskStatusQueries.some(
+                (query) =>
+                    query.isError || query.data?.status === TaskStatus.FAILED,
+            )
+        ) {
+            console.error('One or more OCR pipeline tasks failed');
+            setIsUploading(false);
+            setOcrPipelineTaskIds([]);
+            setProcessingPdfFiles([]);
+            return;
+        }
+
+        // Process all successful tasks
+        if (
+            taskStatusQueries.every(
+                (query) =>
+                    query.isSuccess &&
+                    query.data?.status === TaskStatus.SUCCEEDED,
+            )
+        ) {
+            // from each parsed file, construct a File object
+            const markdownFiles = taskStatusQueries.map((query) => {
+                const file = query.data as TaskResponse;
+                let mdContents = '';
+                for (const chunk of file.output.chunks) {
+                    for (const segment of chunk.segments) {
+                        mdContents += segment.markdown + '\n';
+                    }
+                    mdContents += '\n';
+                }
+                mdContents += '\n';
+                const mdFilename = `${file.output.file_name.split('.pdf')[0]}.md`;
+                const mdBlob = new Blob([mdContents], {
+                    type: 'text/markdown',
+                });
+                return new File([mdBlob], mdFilename);
+            });
+
+            uploadFiles(markdownFiles);
+            console.log('Uploaded files to Gumloop');
+
+            for (let i = 0; i < markdownFiles.length; i++) {
+                const convertedFileName = markdownFiles[i].name;
+                const currentFile = processingPdfFiles[i];
+                updateGumloopName.mutate({
+                    documentId: currentFile.supabaseId,
+                    gumloopName: convertedFileName,
+                    orgId: organizationId,
+                });
+                console.log(
+                    'Updated Gumloop name for file:',
+                    convertedFileName,
+                );
+
+                setSelectedFiles((prev) => ({
+                    ...prev,
+                    [convertedFileName]: currentFile.name,
+                }));
+            }
+
+            setIsUploading(false);
+            setOcrPipelineTaskIds([]);
+            setProcessingPdfFiles([]);
+        }
+    }, [
+        organizationId,
+        processingPdfFiles,
+        setSelectedFiles,
+        taskStatusQueries,
+        updateGumloopName,
+        uploadFiles,
+    ]);
+
+    const [uploadButtonText, setUploadButtonText] = useState('Upload Files');
+
+    useEffect(() => {
+        if (isUploading) {
+            if (ocrPipelineTaskIds.length > 0) {
+                setUploadButtonText('Converting...');
+            } else setUploadButtonText('Uploading...');
+        } else {
+            setUploadButtonText('Upload Files');
+        }
+    }, [isUploading, ocrPipelineTaskIds]);
 
     return (
         <Card className="p-6">
